refactor: domain, service, utils 패키지 생성 후 패키지별 클래스 이동
- parsing 관련 로직을 utils 로 이동
- csv 파일 관련 로직을 utils 로 이동
