+++ src/main/java/domain/Crew.java
@@ -1,3 +1,5 @@
+package domain;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
