+++ src/main/java/service/CrewRegistration.java
@@ -1,0 +1,5 @@
+package service;
+
+public class CrewRegistration {
+    // TODO: CsvReader 와 ParsingUtils 에서 메서드 호출
+}
